requirements.txt
jupyterlab==3.4.4
pyro-ppl==1.8.1
scikit-learn==1.1.2
seaborn==0.11.2
torch==1.12.1
torchvision==0.13.1
